Update README.md
This project is a notepad that helps the user to create plain text files and to write any text into it. Also, it has various features added to it where the user can comfortably save, edit and access all the data that is given by them whenever they require it.